// apps/console/src/__generated__/core/RiskControlsTabFragment.graphql.ts
export type RiskControlsTabFragment$data = {
  readonly controls: {
    readonly edges: ReadonlyArray<{
      readonly node: {
        readonly framework: {
          readonly id: string;
          readonly name: string;
        };
        readonly id: string;
        readonly name: string;
        readonly sectionTitle: string;
      };
    }>;
  };
  readonly id: string;
  readonly " $fragmentType": "RiskControlsTabFragment";
};
export type RiskControlsTabFragment$key = {
  readonly " $data"?: RiskControlsTabFragment$data;
  readonly " $fragmentSpreads": FragmentRefs<"RiskControlsTabFragment">;
};